adda data

+++ react/MWF-8-(2)/src/00.task/todo/ToDoList.jsx
@@ -1,6 +1,7 @@
-import { BadgeCheck, Pencil, Plus, Trash2 } from "lucide-react";
-import React, { useEffect, useState } from "react";
+import { BadgeCheck, Pencil, Plus, Search, Trash2 } from "lucide-react";
+import React, { useEffect, useRef, useState } from "react";
 import { toast } from "react-toastify";
+import Swal from "sweetalert2";
 import {
   Button,
   CardBody,
@@ -19,6 +20,8 @@ export default function TodoProject() {
   const [completedTaskArr, setCompletedTaskArr] = useState([]);
   const [index, setIndex] = useState(null);
   const [pendingSelectedTasks, setPendingSelectedTasks] = useState([]);
+  let [searchTextClick, setSearchTextClick] = useState("");
+  const inputRef = useRef(null);
   const [completedSelectedTasks, setCompletedSelectedTasks] = useState([]);
 
   useEffect(() => {
@@ -46,11 +49,7 @@ export default function TodoProject() {
       toast.warn("Please add task");
     }
   };
-  // console.log("taskArr: ", taskArr);
-
   const addPendingTaskToCompletedTask = (e, index) => {
-    // console.log("e, index", e, index);
-
     setCompletedTaskArr([e, ...completedTaskArr]);
     localStorage.setItem(
       "CompletedTask",
@@ -66,9 +65,45 @@ export default function TodoProject() {
   // console.log("completedTaskArr: ", completedTaskArr);
 
   const deleteHandler = (index) => {
-    let filterData = completedTaskArr.filter((e, i) => i != index);
-    setCompletedTaskArr(filterData);
-    localStorage.setItem("CompletedTask", JSON.stringify(filterData));
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger",
+      },
+      buttonsStyling: false,
+    });
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this task!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          let filterData = completedTaskArr.filter((e, i) => i != index);
+          setCompletedTaskArr(filterData);
+          localStorage.setItem("CompletedTask", JSON.stringify(filterData));
+
+          swalWithBootstrapButtons.fire({
+            title: "Deleted!",
+            text: "Your task has been deleted.",
+            icon: "success",
+          });
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire({
+            title: "Cancelled",
+            text: "Your task is safe.",
+            icon: "error",
+          });
+        }
+      });
   };
 
   const updateHandler = (e, i) => {
@@ -153,14 +188,129 @@ export default function TodoProject() {
   };
 
   const pendingSelectAll = (e) => {
-    if (e.target.checked) {
-      let allIndex = taskArr.map((e, i) => i);
-      setPendingSelectedTasks(allIndex);
+    if (e?.target?.checked) {
+      let allPendingSelectedIndex = taskArr.map((e, i) => i);
+      setPendingSelectedTasks(allPendingSelectedIndex);
     } else {
       setPendingSelectedTasks([]);
     }
   };
 
+  const completedSelectAll = (e) => {
+    if (e?.target?.checked) {
+      let allCompletedSelectedIndex = completedTaskArr.map((e, i) => i);
+      setCompletedSelectedTasks(allCompletedSelectedIndex);
+    } else {
+      setCompletedSelectedTasks([]);
+    }
+  };
+
+  const deletePendingTask = () => {
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger",
+      },
+      buttonsStyling: false,
+    });
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this task!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          let restPendingTasks = taskArr.filter((e, i) => {
+            return !pendingSelectedTasks.includes(i);
+          });
+          setTaskArr(restPendingTasks);
+          localStorage.setItem("PendingTask", JSON.stringify(restPendingTasks));
+          setPendingSelectedTasks([]);
+
+          swalWithBootstrapButtons.fire({
+            title: "Deleted!",
+            text: "Your task has been deleted.",
+            icon: "success",
+          });
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire({
+            title: "Cancelled",
+            text: "Your task is safe.",
+            icon: "error",
+          });
+        }
+      });
+  };
+
+  const deleteCompletedTask = () => {
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger",
+      },
+      buttonsStyling: false,
+    });
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this task!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          let restCompletedTasks = completedTaskArr.filter((e, i) => {
+            return !completedSelectedTasks.includes(i);
+          });
+          setCompletedTaskArr(restCompletedTasks);
+          localStorage.setItem(
+            "CompletedTask",
+            JSON.stringify(restCompletedTasks)
+          );
+          setCompletedSelectedTasks([]);
+
+          swalWithBootstrapButtons.fire({
+            title: "Deleted!",
+            text: "Your task has been deleted.",
+            icon: "success",
+          });
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire({
+            title: "Cancelled",
+            text: "Your task is safe.",
+            icon: "error",
+          });
+        }
+      });
+  };
+
+  const searchData = () => {
+    const allData = JSON.parse(localStorage.getItem("CompletedTask") || "[]");
+
+    setCompletedTaskArr(allData.filter((e) => e.includes(searchTextClick)));
+  };
+
+  const searchHandler = (e) => {
+    setSearchTextClick(e?.target?.value);
+    if (e?.target?.value?.length === 0) {
+      const allData = JSON.parse(localStorage.getItem("CompletedTask") || "[]");
+      setCompletedTaskArr(allData);
+    }
+  };
   return (
     <section className="p-5">
       <div className="container">
@@ -195,8 +345,20 @@ export default function TodoProject() {
             <ReactCard>
               <CardBody>
                 <div className="row">
-                  <div className="col-md-3"></div>
                   <div className="col-md-6">
+                    <FormGroup>
+                      <div className="d-flex align-items-center gap-2">
+                        <Label for="pendingSearch" className="mb-0">
+                          Search:
+                        </Label>
+                        <Input
+                          id="pendingSearch"
+                          placeholder="Search Pending Task..."
+                        />
+                      </div>
+                    </FormGroup>
+                  </div>
+                  <div className="col-md-3">
                     <CardTitle tag="h5" className="text-center">
                       Pending Task
                     </CardTitle>
@@ -261,11 +423,18 @@ export default function TodoProject() {
                   </ListGroup>
                 </ReactCard>
                 <Button
-                  className="w-100 mt-3"
+                  className="w-50 me-2 mt-3"
                   color="primary"
                   onClick={() => addSelectedPendingTaskToCompletedTask()}
                 >
                   Move to Completed Task
+                </Button>
+                <Button
+                  className="ms-2 mt-3"
+                  color="danger"
+                  onClick={() => deletePendingTask()}
+                >
+                  Delete Pending Task
                 </Button>
               </CardBody>
             </ReactCard>
@@ -273,65 +442,100 @@ export default function TodoProject() {
 
           <div className="col-md-6">
             <ReactCard>
-              <CardBody>
-                <div className="row">
-                  <div className="col-md-3"></div>
-                  <div className="col-md-6">
-                    <CardTitle tag="h5" className="text-center">
-                      Completed Task
-                    </CardTitle>
+              {/* <CardBody> */}
+              <div className="row">
+                <div className="col-md-6">
+                  {/* <FormGroup> */}
+                  <div className="d-flex align-items-center gap-2">
+                    <Label for="completedSearch" className="mb-0">
+                      Search:
+                    </Label>
+                    <Input
+                      innerRef={inputRef}
+                      id="completedSearch"
+                      placeholder="Search Completed Task..."
+                      onChange={(e) => searchHandler(e)}
+                    />
+                    <hr />
+                    <Button color="primary" onClick={() => searchData()}>
+                      <Search />
+                    </Button>
                   </div>
-                  <div className="col-md-3 d-flex justify-content-end">
-                    <FormGroup check>
-                      <Input id="completedTaskCheckbox" type="checkbox" />
-                      <Label for="completedTaskCheckbox">Select All</Label>
-                    </FormGroup>
-                  </div>
+                  {/* </FormGroup> */}
                 </div>
-                <ReactCard>
-                  <ListGroup flush>
-                    {completedTaskArr?.map?.((e, i) => {
-                      return (
-                        <ListGroupItem
-                          className="d-flex align-items-center justify-content-between"
-                          key={i}
-                        >
-                          <div className="d-flex align-items-center gap-2">
-                            <Input
-                              type="checkbox"
-                              checked={completedSelectedTasks.includes(i)}
-                              onChange={(e) =>
-                                handleCompletedTaskSelection(
-                                  i,
-                                  e?.target?.checked
-                                )
-                              }
-                            />
-                            <span>
-                              {i + 1}. {e}
-                            </span>
-                          </div>
-                          <div className="d-flex align-items-center gap-2">
-                            <Trash2
-                              role="button"
-                              size={20}
-                              color="#ff0000"
-                              onClick={() => deleteHandler(i)}
-                            />
-                          </div>
-                        </ListGroupItem>
-                      );
-                    })}
-                  </ListGroup>
-                </ReactCard>
-                <Button
-                  className="w-100 mt-3"
-                  color="primary"
-                  onClick={() => addSelectedCompletedTaskToPendingTask()}
-                >
-                  Move to Pending Task
-                </Button>
-              </CardBody>
+                <div className="col-md-3">
+                  {/* <CardTitle tag="h5" className="text-center">
+                      Completed Task
+                    </CardTitle> */}
+                </div>
+                <div className="col-md-3 d-flex justify-content-end">
+                  <FormGroup check>
+                    <Input id="completedTaskCheckbox" type="checkbox" />
+                    <Input
+                      id="completedTaskCheckbox"
+                      type="checkbox"
+                      checked={
+                        completedTaskArr.length > 0 &&
+                        completedTaskArr.length ===
+                          completedSelectedTasks.length
+                      }
+                      onChange={(e) => completedSelectAll(e)}
+                    />
+                    <Label for="completedTaskCheckbox">Select All</Label>
+                  </FormGroup>
+                </div>
+              </div>
+              <ReactCard>
+                <ListGroup flush>
+                  {completedTaskArr?.map?.((e, i) => {
+                    return (
+                      <ListGroupItem
+                        className="d-flex align-items-center justify-content-between"
+                        key={i}
+                      >
+                        <div className="d-flex align-items-center gap-2">
+                          <Input
+                            type="checkbox"
+                            checked={completedSelectedTasks.includes(i)}
+                            onChange={(e) =>
+                              handleCompletedTaskSelection(
+                                i,
+                                e?.target?.checked
+                              )
+                            }
+                          />
+                          <span>
+                            {i + 1}. {e}
+                          </span>
+                        </div>
+                        <div className="d-flex align-items-center gap-2">
+                          <Trash2
+                            role="button"
+                            size={20}
+                            color="#ff0000"
+                            onClick={() => deleteHandler(i)}
+                          />
+                        </div>
+                      </ListGroupItem>
+                    );
+                  })}
+                </ListGroup>
+              </ReactCard>
+              <Button
+                className="w-50 me-2 mt-3"
+                color="primary"
+                onClick={() => addSelectedCompletedTaskToPendingTask()}
+              >
+                Move to Pending Task
+              </Button>
+              <Button
+                className="ms-2 mt-3"
+                color="danger"
+                onClick={() => deleteCompletedTask()}
+              >
+                Delete Completed Task
+              </Button>
+              {/* </CardBody> */}
             </ReactCard>
           </div>
         </div>
